+++ spec/whois/parsers/responses/whois.nic.io/io/status_registered_spec.rb
@@ -33,7 +33,7 @@ describe Whois::Parsers::WhoisNicIo, "status_registered.expected" do
   end
   describe "#domain_id" do
     it do
-      expect { subject.domain_id }.to raise_error(Whois::AttributeNotSupported)
+      expect(subject.domain_id).to eq("D503300000040423931-LRMS")
     end
   end
   describe "#status" do
@@ -53,23 +53,29 @@ describe Whois::Parsers::WhoisNicIo, "status_registered.expected" do
   end
   describe "#created_on" do
     it do
-      expect { subject.created_on }.to raise_error(Whois::AttributeNotSupported)
+      expect(subject.created_on).to be_a(Time)
+      expect(subject.created_on).to eq(Time.parse("2010-05-28 22:09:44 UTC"))
     end
   end
   describe "#updated_on" do
     it do
-      expect { subject.updated_on }.to raise_error(Whois::AttributeNotSupported)
+      expect(subject.updated_on).to be_a(Time)
+      expect(subject.updated_on).to eq(Time.parse("2019-05-06 17:24:58 UTC"))
     end
   end
   describe "#expires_on" do
     it do
       expect(subject.expires_on).to be_a(Time)
-      expect(subject.expires_on).to eq(Time.parse("2014-05-28"))
+      expect(subject.expires_on).to eq(Time.parse("2020-05-28 22:09:44 UTC"))
     end
   end
   describe "#registrar" do
     it do
-      expect { subject.registrar }.to raise_error(Whois::AttributeNotSupported)
+      expect(subject.registrar).to be_a(Whois::Parser::Registrar)
+      expect(subject.registrar.id).to eq("625")
+      expect(subject.registrar.name).to eq("Name.com LLC")
+      expect(subject.registrar.organization).to eq("Name.com LLC")
+      expect(subject.registrar.url).to eq("www.name.com")
     end
   end
   describe "#registrant_contacts" do
@@ -79,14 +85,14 @@ describe Whois::Parsers::WhoisNicIo, "status_registered.expected" do
       expect(subject.registrant_contacts[0]).to be_a(Whois::Parser::Contact)
       expect(subject.registrant_contacts[0].type).to eq(Whois::Parser::Contact::TYPE_REGISTRANT)
       expect(subject.registrant_contacts[0].id).to eq(nil)
-      expect(subject.registrant_contacts[0].name).to eq("Salvatore Sanfilippo")
-      expect(subject.registrant_contacts[0].organization).to eq("Salvatore Sanfilippo")
-      expect(subject.registrant_contacts[0].address).to eq("Via F.Alaimo, 2")
-      expect(subject.registrant_contacts[0].city).to eq("Campobello di Licata (AG")
+      expect(subject.registrant_contacts[0].name).to eq(nil)
+      expect(subject.registrant_contacts[0].organization).to eq("Domain Protection Services, Inc.")
+      expect(subject.registrant_contacts[0].address).to eq(nil)
+      expect(subject.registrant_contacts[0].city).to eq(nil)
       expect(subject.registrant_contacts[0].zip).to eq(nil)
-      expect(subject.registrant_contacts[0].state).to eq(".")
-      expect(subject.registrant_contacts[0].country).to eq("IT")
-      expect(subject.registrant_contacts[0].country_code).to eq(nil)
+      expect(subject.registrant_contacts[0].state).to eq("CO")
+      expect(subject.registrant_contacts[0].country).to eq(nil)
+      expect(subject.registrant_contacts[0].country_code).to eq("US")
       expect(subject.registrant_contacts[0].phone).to eq(nil)
       expect(subject.registrant_contacts[0].fax).to eq(nil)
       expect(subject.registrant_contacts[0].email).to eq(nil)
@@ -109,13 +115,13 @@ describe Whois::Parsers::WhoisNicIo, "status_registered.expected" do
       expect(subject.nameservers).to be_a(Array)
       expect(subject.nameservers.size).to eq(4)
       expect(subject.nameservers[0]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[0].name).to eq("ns1.iwantmyname.net")
+      expect(subject.nameservers[0].name).to eq("ns2kqz.name.com")
       expect(subject.nameservers[1]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[1].name).to eq("ns2.iwantmyname.net")
+      expect(subject.nameservers[1].name).to eq("ns1djs.name.com")
       expect(subject.nameservers[2]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[2].name).to eq("ns3.iwantmyname.net")
+      expect(subject.nameservers[2].name).to eq("ns4bht.name.com")
       expect(subject.nameservers[3]).to be_a(Whois::Parser::Nameserver)
-      expect(subject.nameservers[3].name).to eq("ns4.iwantmyname.net")
+      expect(subject.nameservers[3].name).to eq("ns3ckl.name.com")
     end
   end
   describe "#reserved?" do