+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 
-#### master
+#### Release 1.2.0
 
 - CHANGED: Updated .ORG parser to the latest response (GH-98, GH-97). [Thanks @talarini]
 - CHANGED: Updated .IO, .AC, .SH, .TM parsers to the latest response.